+++ R/fill_yaml.R
@@ -25,6 +25,8 @@ fill_yaml <- function(yml,
                       docker_token,
                       cache_version,
                       enable_act){
+  # devoptera::args2vars(use_workflow)
+  
   #### name ####
   yml$name <- name
   #### on ####
@@ -49,23 +51,14 @@ fill_yaml <- function(yml,
     }  
   }
   #### Supply variables as "with:" or "env:" ####
-  with2$run_bioccheck <- run_bioccheck
-  with2$run_rcmdcheck <- run_rcmdcheck
-  with2$as_cran <- as_cran
-  with2$run_vignettes <- run_vignettes
-  with2$has_testthat <- has_testthat
-  with2$run_covr <- run_covr
-  with2$run_pkgdown <- run_pkgdown
-  with2$has_runit <- has_runit 
-  with2$has_latex <- has_latex 
-  with2$tinytex_installer <- tinytex_installer
-  with2$tinytex_version <- tinytex_version
-  with2$GITHUB_TOKEN <- github_token 
-  with2$run_docker <- run_docker 
-  with2$docker_user <- docker_user 
-  with2$docker_org <- docker_org 
-  with2$DOCKER_TOKEN <- docker_token 
-  with2$cache_version <- cache_version  
+  nonarg_list <- c("yml","name","template","tag","on","branches","runners")
+  args_list <- names(formals(fill_yaml))
+  args_list <- args_list[!args_list %in% nonarg_list]
+  omit_list <- c("tinytex_installer","tinytex_version")
+  for(a in args_list){
+    nm <- if(a %in% c("github_token","docker_token")) toupper(a) else a
+    with2[nm] <- if(a %in% omit_list) omit_if_default(arg = a) else get(a)
+  }   
   #### static workflow vs. action ####
   if(template=="rworkflows_static"){
     yml$env <- with2